requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0
rapidfuzz>=3.6
pyarabic>=0.6.15
jellyfish>=1.0
sentence-transformers>=2.6
faiss-cpu>=1.7.4
numpy>=1.24
anthropic>=0.40
openai>=1.30
deepgram-sdk>=3.7
pytest>=8.0
httpx>=0.27
